+++ ChangeLog.md
@@ -6,6 +6,10 @@ XP Framework Core ChangeLog
 ### Heads up!
 
 * Removed deprecated `lang.Enum::membersOf()` method - (@thekid)
+* Removed deprecated methods in util.Date:
+  - mktime() - replaced by create()
+  - fromString() - handled by regular constructor
+  (@thekid)
 * Removed deprecated add*() methods in util.TimeSpan class - (@thekid)
 * Removed support for deprecated multi-value annotations - (@thekid)
 * Removed deprecated lang.ChainedException - (@thekid)